refactor(auth): adjust file structure and naming
