+++ auth/src/utils/email-sender/__tests__/email-sender.test.ts
@@ -1,9 +1,9 @@
-import { EmailSender } from '../index';
+import { EmailSender } from '../../index';
 import {
   MockEmailApi,
   mockSendSignupVerificationEmail,
   mockSendEmail
-} from '../../test-utils/mock-email-api';
+} from '../../../test-utils/mock-email-api';
 
 it('should throw an error when sending an email if the EmailSender is deactivated', async () => {
   const emailSender = EmailSender.getInstance();
