+++ auth/src/utils/email-sender/index.ts
@@ -1,0 +1,1 @@
+export { default as EmailSender } from './email-sender';
