+++ auth/src/utils/index.ts
@@ -1,2 +1,2 @@
-export { default as PasswordHash } from './password-hash';
-export { default as EmailSender } from './email-sender';
+export * from './password-hash';
+export * from './email-sender';
