+++ auth/src/utils/password-hash/index.ts
@@ -1,0 +1,1 @@
+export { default as PasswordHash } from './password-hash';
